move the non-API routes
